Reproduce issue in spec
Moved other pagination-related specs to new location, to better organise
specs as well as to share the helper `expect_to_appear_in_order`.
