+++ spec/features/index_page_spec.rb
@@ -68,72 +68,6 @@ describe "customer index page" do
       expect(page).to have_table_header(custom_label)
     end
   end
-
-  it "paginates records based on a constant" do
-    customers = create_list(:customer, 2)
-
-    visit admin_customers_path(per_page: 1)
-
-    expect(page).not_to have_content(customers.last.name)
-    click_on "Next"
-    expect(page).to have_content(customers.last.name)
-  end
-
-  describe "sorting" do
-    def expect_to_appear_in_order(*elements)
-      positions = elements.map { |e| page.body.index(e) }
-      expect(positions).to eq(positions.sort)
-    end
-
-    it "allows sorting by columns" do
-      create(:customer, name: "unique name two")
-      create(:customer, name: "unique name one")
-
-      visit admin_customers_path
-      click_on "Name"
-
-      expect_to_appear_in_order("unique name one", "unique name two")
-    end
-
-    it "allows clicking through after sorting", :js do
-      customer = create(:customer)
-      create(:order, customer: customer)
-
-      visit admin_customers_path
-      click_on "Name"
-      find("[data-url]").click
-      expect(page).to have_header("Show #{customer.name}")
-    end
-
-    it "allows reverse sorting" do
-      create(:customer, name: "unique name one")
-      create(:customer, name: "unique name two")
-
-      visit admin_customers_path
-      2.times { click_on "Name" }
-
-      expect_to_appear_in_order("unique name two", "unique name one")
-    end
-
-    it "toggles the order" do
-      create(:customer, name: "unique name one")
-      create(:customer, name: "unique name two")
-
-      visit admin_customers_path
-      3.times { click_on "Name" }
-
-      expect_to_appear_in_order("unique name one", "unique name two")
-    end
-
-    it "preserves search" do
-      query = "bar@baz.com"
-
-      visit admin_customers_path(search: query)
-      click_on "Name"
-
-      expect(find(search_input_selector).value).to eq(query)
-    end
-  end
 end
 
 describe "search input" do
